pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "faelbochi", email = ""},
]
dependencies = [
    "django>=4.2.2",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.2.2",
    "setuptools>=68.0.0",
    "drf-spectacular>=0.26.3",
    "django-cors-headers>=4.1.0",
    "django-extensions>=3.2.3",
    "pygraphviz>=1.11",
    "pydotplus>=2.0.2",
]
requires-python = ">=3.11"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"